+++ frontend/src/Components/Header.tsx
@@ -1,0 +1,8 @@
+import {Box} from '@chakra-ui/react';
+
+const Header = () => {
+    return(
+        <Box><h1>Header</h1></Box>
+    );
+}
+export default Header;
