+++ frontend/src/Pages/LogIn.tsx
@@ -1,0 +1,10 @@
+import {Box} from '@chakra-ui/react';
+
+const LogIn = () => {
+    return(
+        <Box>
+            <h1>LogIn</h1>
+        </Box>
+    );
+}
+export default LogIn;
